Extract RazorEditService.CSharpMember
